Add 'if' to grammar (ambiguously)
If we implement 'if' naively, it results in an ambiguous grammar.

lalrpop explains it perfectly (just run it), but here's the example
without the ASCII art:

    if (cond)
        if (cond2)
            stmt1;
        else
            stmt2;

...versus...

    if (cond)
        if (cond2)
            stmt1;
    else
        stmt2;

...which is it?

See https://github.com/lalrpop/lalrpop/issues/386, which suggests using
conditional macros to fix this. But there's a comment that points out
it's flawed.

It points to https://www.cs.cornell.edu/andru/javaspec/19.doc.html for
the solution.

For this commit, I'm going to leave the grammar ambiguous, as a learning
opportunity.