Fix scoping for blocks
In order that 'var' doesn't escape from a nested block, we need to
create a nested (enclosed) environment.

With a little more effort, this is a potential fix to the bug as
explained in 11.1 in the book.

Specifically, rather than mess around with the resolver stuff, we could
just create a new nested environment for every 'var' declaration (this
is a suggested alternative implementation) and we'd fix the problem.

Oh, and I also added a bunch more scope examples.

Still need to figure out how to run them automatically and assert the
output, though.